requirements.txt
# Core dependencies
requests>=2.31.0
aiohttp>=3.9.0
asyncio>=3.4.3
python-dotenv>=1.0.0

# Data processing
pandas>=2.1.0
jsonlines>=4.0.0

# Rate limiting and retry
tenacity>=8.2.3
ratelimit>=2.2.1

# Logging and monitoring
tqdm>=4.66.0
colorlog>=6.7.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
